+++ boadata/gui/qt/data_object_window.py
@@ -9,9 +9,11 @@ class DataObjectWindow(QtGui.QMainWindow):
 
         self.tabWidget = QtGui.QTabWidget(self)
         for view in View.registered_views:
+            # print(data_object)
             if view.accepts(data_object):
                 self.tabWidget.addTab(view(data_object).create_widget(), view.title)
             else:
+                # print("View {0} not supported.".format(view.title))
                 pass
                 # self.tabWidget.addTab(QtGui.QLabel("Not supported"), view.title)